+++ src/components/Bubble/Bubble.tsx
@@ -2,11 +2,13 @@ import React from "react";
 
 import styles from "./Bubble.module.scss";
 
-interface BubbleProps {}
+interface BubbleProps {
+  animate: "flyAway" | "orbital";
+}
 
-const Bubble: React.FC<BubbleProps> = () => {
+const Bubble: React.FC<BubbleProps> = ({ animate }) => {
   return (
-    <div className={styles.bubble}>
+    <div className={`${styles.bubble} ${animate === "orbital" ? styles.orbital : styles.flyAway}`}>
       <span></span>
       <span></span>
     </div>
