+++ src/components/ReviewBubble/ReviewBubble.tsx
@@ -1,15 +1,26 @@
-import React from "react";
+import React, { useState } from "react";
+import { generateBubbles } from "@/helpers/generateBubbles";
 
 import styles from "./ReviewBubble.module.scss";
+import { Bubble } from "../Bubble";
 
 interface ReviewBubbleProps {}
 
 const ReviewBubble: React.FC<ReviewBubbleProps> = () => {
+  const [bubbles, setBubbles] = useState<{ [key: string]: boolean }>({});
+
+  const generateBubblesHandleClick = () => {
+    generateBubbles(setBubbles);
+  };
+
   return (
-    <div className={styles.reviewBubble}>
-      <span></span>
-      <span></span>
-    </div>
+    <>
+      <div className={styles.reviewBubble} onClick={generateBubblesHandleClick}>
+        <span></span>
+        <span></span>
+      </div>
+      {Object.entries(bubbles).map(([bubbleId, isVisible]) => isVisible && <Bubble key={bubbleId} animate="orbital" />)}
+    </>
   );
 };
