+++ src/components/RoundElement/RoundElement.tsx
@@ -2,9 +2,9 @@
 import React, { forwardRef, HTMLAttributes, useState } from "react";
 
 import { motion } from "framer-motion";
-import { nanoid } from "nanoid";
 import { Bubble } from "../Bubble";
 import styles from "./RoundElement.module.scss";
+import { generateBubbles } from "@/helpers/generateBubbles";
 
 type TSize = "s" | "m" | "l" | "xl";
 
@@ -31,27 +31,18 @@ const RoundElement = forwardRef<HTMLDivElement, RoundElementProps>(({ text, size
 
   const [bubbles, setBubbles] = useState<{ [key: string]: boolean }>({});
 
-  const handleButtonClick = () => {
-    const bubbleId = nanoid();
-    setBubbles((prev) => ({ ...prev, [bubbleId]: true }));
-
-    setTimeout(() => {
-      setBubbles((prev) => {
-        const newBubbles = { ...prev };
-        delete newBubbles[bubbleId];
-        return newBubbles;
-      });
-    }, 6000);
+  const generateBubblesHandleClick = () => {
+    generateBubbles(setBubbles);
   };
 
   return (
     <div className={`${styles.roundElement} ${roundSize(size)}`} {...props} ref={ref}>
-      <button className={styles.roundElementButton} onClick={handleButtonClick}>
+      <button className={styles.roundElementButton} onClick={generateBubblesHandleClick}>
         <span className={styles.text}>{text}</span>
         <span className={styles.shadowLeft}></span>
         <span className={styles.shadowRight}></span>
       </button>
-      {Object.entries(bubbles).map(([bubbleId, isVisible]) => isVisible && <Bubble key={bubbleId} />)}
+      {Object.entries(bubbles).map(([bubbleId, isVisible]) => isVisible && <Bubble key={bubbleId} animate="flyAway" />)}
     </div>
   );
 });
